+++ src/py42/_internal/client_factories.py
@@ -10,7 +10,7 @@ from py42.clients.detectionlists.departing_employee import DepartingEmployeeClie
 from py42.clients.detectionlists.high_risk_employee import HighRiskEmployeeClient
 from py42.clients.detectionlists.detection_list_user_client import DetectionListUserClient
 from py42.clients.file_event import FileEventClient
-from py42.sdk.exceptions import Py42FeatureUnavailableError, Py42SessionInitializationError
+from py42.exceptions import Py42FeatureUnavailableError, Py42SessionInitializationError
 
 
 class AuthorityClientFactory(object):