+++ src/py42/_internal/storage_session_manager.py
@@ -3,7 +3,7 @@ from threading import Lock
 from requests import HTTPError
 
 from py42._internal.compat import str
-from py42.sdk.exceptions import Py42StorageSessionInitializationError
+from py42.exceptions import Py42StorageSessionInitializationError
 
 
 class StorageSessionManager(object):