+++ src/py42/clients/legalhold.py
@@ -2,7 +2,7 @@ import json
 
 from py42.clients import BaseClient
 from py42.clients.util import get_all_pages
-from py42.sdk.exceptions import Py42Error
+from py42.exceptions import Py42Error
 
 
 def _active_state_map(active):
@@ -35,7 +35,7 @@ class LegalHoldClient(BaseClient):
                 None (where the server-default backup set is used).
 
         Returns:
-            :class:`py42.sdk.response.Py42Response`
+            :class:`py42.response.Py42Response`
         """
         uri = u"/api/v4/legal-hold-policy/create"
         data = {u"name": name, u"policy": policy}
@@ -54,7 +54,7 @@ class LegalHoldClient(BaseClient):
             hold_ext_ref (str, optional): Optional external reference information. Defaults to None.
 
         Returns:
-            :class:`py42.sdk.response.Py42Response`
+            :class:`py42.response.Py42Response`
         """
         uri = u"/api/LegalHold"
         data = {
@@ -74,7 +74,7 @@ class LegalHoldClient(BaseClient):
             legal_hold_policy_uid (str): The identifier of the Preservation Policy.
 
         Returns:
-            :class:`py42.sdk.response.Py42Response`: A response containing the Policy.
+            :class:`py42.response.Py42Response`: A response containing the Policy.
         """
         uri = u"/api/v4/legal-hold-policy/view"
         params = {u"legalHoldPolicyUid": legal_hold_policy_uid}
@@ -85,7 +85,7 @@ class LegalHoldClient(BaseClient):
         `V4 REST Documentation <https://console.us.code42.com/swagger/#/legal-hold-policy/LegalHoldPolicy_List>`__
 
         Returns:
-            :class:`py42.sdk.response.Py42Response`: A response containing the list of Policies.
+            :class:`py42.response.Py42Response`: A response containing the list of Policies.
         """
         uri = u"/api/v4/legal-hold-policy/list"
         return self._session.get(uri)
@@ -98,7 +98,7 @@ class LegalHoldClient(BaseClient):
             legal_hold_uid (str): The identifier of the Legal Hold Matter.
 
         Returns:
-            :class:`py42.sdk.response.Py42Response`: A response containing the Matter.
+            :class:`py42.response.Py42Response`: A response containing the Matter.
         """
         uri = u"/api/LegalHold/{0}".format(legal_hold_uid)
         return self._session.get(uri)
@@ -140,7 +140,7 @@ class LegalHoldClient(BaseClient):
                 Defaults to None.
 
         Returns:
-            generator: An object that iterates over :class:`py42.sdk.response.Py42Response` objects
+            generator: An object that iterates over :class:`py42.response.Py42Response` objects
             that each contain a page of Legal Hold Matters.
         """
         return get_all_pages(
@@ -197,7 +197,7 @@ class LegalHoldClient(BaseClient):
                 None returns all LegalHoldMemberships regardless of state. Defaults to True.
 
         Returns:
-            generator: An object that iterates over :class:`py42.sdk.response.Py42Response` objects
+            generator: An object that iterates over :class:`py42.response.Py42Response` objects
             that each contain a page of LegalHoldMembership objects.
         """
         return get_all_pages(
@@ -218,7 +218,7 @@ class LegalHoldClient(BaseClient):
             legal_hold_uid (str): The identifier of the Legal Hold Matter.
 
         Returns:
-            :class:`py42.sdk.response.Py42Response`
+            :class:`py42.response.Py42Response`
         """
         uri = u"/api/LegalHoldMembership"
         data = {u"legalHoldUid": legal_hold_uid, u"userUid": user_uid}
@@ -233,7 +233,7 @@ class LegalHoldClient(BaseClient):
                 the Custodian to Matter relationship.
 
         Returns:
-            :class:`py42.sdk.response.Py42Response`
+            :class:`py42.response.Py42Response`
         """
         uri = u"/api/LegalHoldMembershipDeactivation"
         data = {u"legalHoldMembershipUid": legal_hold_membership_uid}
@@ -247,7 +247,7 @@ class LegalHoldClient(BaseClient):
             legal_hold_uid (str): The identifier of the Legal Hold Matter.
 
         Returns:
-            :class:`py42.sdk.response.Py42Response`
+            :class:`py42.response.Py42Response`
         """
         uri = u"/api/v4/legal-hold-deactivation/update"
         data = {u"legalHoldUid": legal_hold_uid}
@@ -261,7 +261,7 @@ class LegalHoldClient(BaseClient):
             legal_hold_uid (str): The identifier of the Legal Hold Matter.
 
         Returns:
-            :class:`py42.sdk.response.Py42Response`
+            :class:`py42.response.Py42Response`
         """
         uri = u"/api/LegalHoldReactivation/{0}".format(legal_hold_uid)
         return self._session.put(uri)