+++ tests/_internal/clients/storage/test_archive.py
@@ -4,7 +4,7 @@ import pytest
 
 from py42._internal.clients.storage import StorageArchiveClient
 from py42._internal.session import Py42Session
-from py42.sdk.response import Py42Response
+from py42.response import Py42Response
 
 DATA_KEYWORD = "data"
 JSON_KEYWORD = "json"