+++ tests/clients/test_legalhold.py
@@ -3,7 +3,7 @@ from requests import Response
 
 import py42
 from py42.clients.legalhold import LegalHoldClient
-from py42.sdk.response import Py42Response
+from py42.response import Py42Response
 
 LEGAL_HOLD_URI = "/api/LegalHold"
 
@@ -74,7 +74,7 @@ class TestLegalHoldClient(object):
     def test_get_all_matters_calls_get_expected_number_of_times(
         self, mock_session, mock_get_all_matters_response, mock_get_all_matters_empty_response
     ):
-        py42.sdk.settings.items_per_page = 1
+        py42.settings.items_per_page = 1
         client = LegalHoldClient(mock_session)
         mock_session.get.side_effect = [
             mock_get_all_matters_response,
@@ -83,7 +83,7 @@ class TestLegalHoldClient(object):
         ]
         for _ in client.get_all_matters():
             pass
-        py42.sdk.settings.items_per_page = 1000
+        py42.settings.items_per_page = 1000
         assert mock_session.get.call_count == 3
 
     def test_get_all_matter_custodians_calls_get_expected_number_of_times(
@@ -92,7 +92,7 @@ class TestLegalHoldClient(object):
         mock_get_all_matter_custodians_response,
         mock_get_all_matter_custodians_empty_response,
     ):
-        py42.sdk.settings.items_per_page = 1
+        py42.settings.items_per_page = 1
         client = LegalHoldClient(mock_session)
         mock_session.get.side_effect = [
             mock_get_all_matter_custodians_response,
@@ -101,5 +101,5 @@ class TestLegalHoldClient(object):
         ]
         for _ in client.get_all_matter_custodians():
             pass
-        py42.sdk.settings.items_per_page = 1000
+        py42.settings.items_per_page = 1000
         assert mock_session.get.call_count == 3