+++ tests/sdk/test_util.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-import py42.sdk.util as util
+import py42.util as util
 
 
 def test_convert_timestamp_to_str_returns_expected_str():